Refactor some dependencies can be made optional.
Main motivating factor was the calibrator needed scikit-learn, in particular a very old version. This has been causing issues. So we make a new calibrator class which mimics the behavior and can be pickled/unpickled with no dependencies (besides itself); the results are exactly the same. Tweak some other imports as well for similar purpose. Many files needed a bit of modification as a result.

* Modify setup.cfg so we don't need to install sklearn unless training.

* Leave the old config in for now, will remove it next version.

* Fix spelling of "calibrator"